se arreglo mensajes, para que esten en el config

+++ app/Http/Controllers/CategoryController.php
@@ -40,13 +40,18 @@ class CategoryController extends Controller
 
         Category::create($request->all());
 
-        Alert::success('Se creo el registro.', 'Excelente!');
+        Alert::success(config('guiar.msj.success_create_record'), config('success_title'));
         return redirect()->route('categories.index');
     }
 
     public function edit($uuid)
     {
         $category = Category::where('uuid',$uuid)->first();
+
+        if(!$category){
+            Alert::error(config('guiar.msj.fail_record_not_found'), config('fail_title'));
+            return redirect()->route('categories.index');
+        }
 
         $data=[
             'category' => $category
@@ -60,13 +65,18 @@ class CategoryController extends Controller
 
         $category = Category::where('uuid',$uuid)->first();
 
+        if(!$category){
+            Alert::error(config('guiar.msj.fail_record_not_found'), config('fail_title'));
+            return redirect()->route('categories.index');
+        }
+
         $this->validate($request, [
             'name' => 'required|unique:categories,id,'.$category->id
         ]);
 
         $category->update($request->all());
 
-        Alert::success('Se actualizo el registro.', 'Excelente!');
+        Alert::success(config('guiar.msj.success_update_record'), config('success_title'));
         return redirect()->route('categories.index');
     }
 
@@ -74,9 +84,14 @@ class CategoryController extends Controller
     {
         $category = Category::where('uuid',$uuid)->first();
 
+        if(!$category){
+            Alert::error(config('guiar.msj.fail_record_not_found'), config('fail_title'));
+            return redirect()->route('categories.index');
+        }
+
         $category->delete();
 
-        Alert::success('Se ha eliminado el registro.', 'Excelente!');
+        Alert::success(config('guiar.msj.success_delete_record'), config('success_title'));
         return redirect()->route('categories.index');
     }
